Removed role by react code Commented out some things that may end up being useful if reused Any remaining democracy code should have been removed by now Other small changes for QOL

+++ run.py
@@ -6,6 +6,7 @@ import os
 import sys
 import time
 import yaml
+import shutil
 import random
 from datetime import datetime
 from timeit import default_timer as timer
@@ -35,38 +36,65 @@ protected_example_addresses = [
 ]
 
 config = configparser.ConfigParser()
-config.read(os.path.abspath("./configs/config.ini"))
-admin_role_names = config["Credentials"]["admin_role_names"]
-vote_right_role = config["Credentials"]["vote_rights_role"]
+
+try:
+    config.read(os.path.abspath("./configs/config.ini"))
+except FileNotFoundError:
+    try:
+        shutil.copyfile("./configs/default_config.ini", "./configs/config.ini")
+        print("You need to set up the config file correctly.")
+    except shutil.Error:
+        print("Something is wrong with the default config file or the config folder.")
+        time.sleep(4)
+
+    sys.exit()
+
+admin_role_names = config["Credentials"]["admin_roles"]
 bot_token = config["Credentials"]["bot_token"]
+colour_roles = config["Credentials"]["allowed_colour_requesters"]
 use_drive_for_backup = config["Credentials"]["use_drive_for_backup"]
-owner_id = config["Credentials"]["owner_id"]
+owner_id = int(config["Credentials"]["owner_id"])
 mute_role_name = config["Misc"]["mute_role_name"]
 student_role_name = config["Misc"]["student_role_name"]
+exclusion_colours = config["Misc"]["exclusion_colours"]
+exclusion_range = config["Misc"]["exclusion_side_length"]
 delete_messages_after = config["Misc"]["delete_messages_after"]
 listen_channels = config["Misc"]["listen_channels"]
-announce_channel_config = config["Misc"]["announcement_channel_id"]
-cmd_prefix = "££"  # COMMAND PREFIX IS HERE FOR EDITING PURPOSES, UNICODE WAS BEING A FUCK SO THAT'S WHY IT'S HERE
+role_channel_id = config["Misc"]["role_channel_id"]
+
+cmd_prefix = "CPS."  # COMMAND PREFIX IS HERE FOR EDITING PURPOSES, UNICODE WAS BEING A FUCK SO THAT'S WHY IT'S HERE
 
 admin_role_list = admin_role_names.split(",")
+colour_request_list = colour_roles.split(",")
+extra_exclusion_colours = exclusion_colours.split(",")
 listen_channels_list = listen_channels.split(",")
 
 if admin_role_list[0] == "":
     print("You must specify administrative roles for this bot.")
     sys.exit("No admin roles defined.")
 
+if colour_request_list[0] == "":
+    colour_request_list = admin_role_list
+
 if listen_channels_list[0] == "":
     listen_channels_list.pop(0)
 
-try:
-    use_drive_for_backup = bool(use_drive_for_backup)
-except TypeError:
-    print("You must enter a boolean type for use_drive_for_backup, remember to capitalise True/False.")
+if exclusion_range == "":
+    exclusion_range = 20
+else:
+    exclusion_range = int(exclusion_range)
 
-vote_file_queue = []
-vote_types = ["freddie_style_vote", "fpbtp_style_vote"]
+if extra_exclusion_colours[0] == "":
+    extra_exclusion_colours = []
+
+if use_drive_for_backup == "True":
+    use_drive_for_backup = True
+else:
+    use_drive_for_backup = False
 
 bot = commands.Bot(command_prefix=cmd_prefix, pm_help=True)
+
+watching = discord.Activity(type=discord.ActivityType.watching, name="you")
 
 
 async def auth_with_the_gargle():
@@ -150,6 +178,11 @@ async def search_for_file_drive(file_data, query, make_if_missing=False):
 
 
 @bot.event
+async def on_connect():
+    await bot.change_presence(activity=watching)
+
+
+@bot.event
 async def on_ready():
     logger.debug("Start of on_ready()")
 
@@ -165,12 +198,8 @@ async def on_message(msg):
     if msg.author.bot:
         return
 
-    author_obj = msg.author
-    author_is_admin = False
-    for admin_role in author_obj.roles:
-        if admin_role.name in admin_role_list:
-            author_is_admin = True
-            break
+    if msg.guild is None:
+        return
 
     message_protected = False
     for protected_address in protected_example_addresses:
@@ -178,48 +207,58 @@ async def on_message(msg):
             message_protected = True
             break
 
-    if not message_protected:
-        if "@soton.ac.uk" in msg.content and not author_is_admin:
+    author_obj = msg.author
+    author_is_admin = False
+    for admin_role in author_obj.roles:
+        if admin_role.name in admin_role_list:
+            author_is_admin = True
+
+    if not message_protected and not author_is_admin:
+        if "@soton.ac.uk" in msg.content:
             await msg.channel.send(f"{msg.author.mention}, Please do not send messages in public channels that contain "
                                    f"your email. If you are attempting to verify with SVGEBot, use direct messages and "
                                    f"send:\n\n`!email your_email@soton.ac.uk` then follow instructions sent to your "
                                    f"university inbox. \n\nIf you believe your message was deleted in error, contact "
-                                   f"`@Freddie (Pytato)` with the datetime of your message.",
+                                   f"<@{str(owner_id)}> with the datetime of your message.",
                                    delete_after=delete_messages_after)
+            await msg.author.send("Below is the message this bot has just deleted:\n\n" + msg.content)
             await msg.delete()
             logger.info(f"Auto-deleted message sent by {msg.author.display_name}, in {msg.channel.name}.")
+            return
 
     if listen_channels_list:
-        if msg.channel.id not in listen_channels_list and msg.guild is not None:
+        if msg.channel.id not in listen_channels_list:
             return
 
     if not msg.content.startswith(cmd_prefix):
         return
 
-    logger.debug("Message received from {0.name}#{0.discriminator} in {1.guild}, #{1.channel.name}."
-                 .format(msg.author, msg))
-
-    if msg.split(" ")[0] != f"{cmd_prefix}vote_for":
-        if not author_is_admin:
-            return
-
     logger.info(f'Command sent by "{msg.author.name}#{msg.author.discriminator}": "{msg.content}."')
 
-    await bot.process_commands(msg)
-
-    await msg.delete()
+    try:
+        msg_preserved = msg
+        await msg.delete()
+        await bot.process_commands(msg_preserved)
+    except discord.ext.commands.CommandNotFound:
+        pass
+    except discord.ext.commands.CheckFailure:
+        logger.warning(f'User: "{msg.author.name}#{msg.author.discriminator}" issued command "{msg.content}". '
+                       f'which failed command checks.')
 
 
 @bot.command()
+@commands.has_any_role(*admin_role_list)
 async def shutdown(ctx):
     """Shuts the bot down as gracefully as possible."""
     await ctx.send(":wave:", delete_after=1)
     await asyncio.sleep(3)
     await bot.logout()
+    await asyncio.sleep(2)
     sys.exit(0)
 
 
-@bot.command(name="warn")
+@bot.command()
+@commands.has_any_role(*admin_role_list)
 async def warn(ctx, target_user_mention, search_depth: int, delete_found_messages: bool, should_mute: bool, *,
                reason: str):
     """This command is used to warn users for breaching the rules, the bot will automatically apply the mute-role
@@ -380,13 +419,15 @@ async def warn(ctx, target_user_mention, search_depth: int, delete_found_message
     logger.debug(f"Finished running warn command in {timer() - warn_command_start}s.")
 
 
-@bot.command
+@bot.command()
+@commands.has_any_role(*admin_role_list)
 async def get_emote_id(ctx, emote):
     """Returns the ID for a given emote, by name or the emote sent in a message.
 
     Args:
         - emote: either emote name or object sent in message, if *, command returns a list of all emotes in DMs,
-        if *ani, returns a list of all animated emotes."""
+        if *ani, returns a list of all animated emotes.
+    """
 
     if emote != "*" and emote != "*ani":
         try:
@@ -410,12 +451,15 @@ async def get_emote_id(ctx, emote):
     await ctx.author.send(f"List of {emote} emotes: \n\n{emote_list_str}", delete_after=delete_messages_after)
 
 
-@bot.command
+'''
+@bot.command()
+@commands.has_any_role(tuple(admin_role_list))
 async def get_role_id(ctx, *, role: str):
     """Returns the Role ID for a given role.
 
     Args:
-        - role: role name existing in the server of server invocation."""
+        - role: role name existing in the server of invocation.
+    """
 
     try:
         role_obj = await commands.RoleConverter().convert(ctx, role)
@@ -424,140 +468,128 @@ async def get_role_id(ctx, *, role: str):
         await ctx.send("Failed to convert role to ID.", delete_after=delete_messages_after)
         return
 
-    await ctx.author.send(f"{role_obj.name} | {str(role_obj.id)}")
+    await ctx.author.send(f"{role_obj.name} | {str(role_obj.id)}", delete_after=delete_messages_after)
+'''
 
 
-@bot.command
-async def create_react_roles(ctx, *, emote_role_dict: dict):
-    """Generates a react message for role allocation
+@bot.command(name="colourme")
+@commands.has_any_role(*colour_request_list)
+async def colour_me(ctx, colour_hex: str):
+    """Gives the command invoker a custom colour role if they satisfy given conditions.
 
-    Args:
-        - emote_role_dict: A comma separated dictionary of emote IDs and their associated role IDs"""
+    If colour_hex is given as remove, the bot will remove the colour role and exit the
+    operation.
+    """
 
-    '''
-        for keys, values in emote_role_dict.items():
-            pass
-    '''
+    # Preprocess the colour
+    if colour_hex.lower() == "remove":
+        for role in ctx.author.roles:
+            if "CPS[0x" in role.name:
+                await ctx.author.remove_roles(role, reason="User requested colour role removal.")
 
-    return
-
-
-@bot.command(pass_context=True)
-async def start_vote(ctx, vote_type_new, vote_name, *, candidate_list: str):
-    is_admin = False
-    for admin_role in admin_role_list:
-        if await commands.RoleConverter().convert(ctx, admin_role) in ctx.author.roles:
-            is_admin = True
-            break
-
-    if not is_admin:
-        logger.warning("User: {} attempted to run a vote but doesn't have correct permissions!"
-                       .format(ctx.author))
+        await asyncio.sleep(0.5)
+        for role in ctx.guild.roles:
+            if "CPS[0x" in role.name:
+                if not role.members:
+                    await role.delete(reason="Automatic custom colour deletion when unused.")
         return
 
-    candidate_list = candidate_list.split(",")
+    if len(colour_hex) > 6:
+        await ctx.send("The colour string requested is invalid.", delete_after=delete_messages_after)
+        return
+    colour_hex_split = [colour_hex[0:2], colour_hex[2:4], colour_hex[4:6]]
+    colour_dec_split = []
+    for colour in colour_hex_split:
+        try:
+            colour_dec = int(colour, 16)
+        except ValueError:
+            return
+        if not (0 < colour_dec < 255):
+            await ctx.message(f"The colour: {colour_hex} sits outside of permitted ranges.",
+                              delete_after=delete_messages_after)
+            return
+        colour_dec_split.append(colour_dec)
 
-    logger.debug("Successfully finished parsing list of candidates.\n" + str(candidate_list))
+    exclusion_cube_origins = []
 
-    announce_message = f"This is a member only vote for {vote_name}" \
-                       f"Below will be the valid candidates for you to vote for, below that the voting format and \n\n" \
-                       f"__Eligible Candidates__:\n"
+    # Set up exclusion zones for colours
+    for admin_role_name in admin_role_list:
+        # Let's first gather all the admin role
+        try:
+            admin_role = await commands.RoleConverter().convert(ctx, admin_role_name)
+            # Now find its colour and add it to the list of exclusion origins
+            admin_role_colour = admin_role.colour.to_rgb()
+            exclusion_cube_origins.append(list(admin_role_colour))
+        except discord.ext.commands.errors.BadArgument:
+            logger.info("Admin role defined in config not found in guild.")
 
-    for candidate in candidate_list:
-        announce_message = announce_message+f"- {candidate}\n"
+    for extra_exclusion_colour in extra_exclusion_colours:
+        hex_exclusion_colour_split = [extra_exclusion_colour[0:2],
+                                      extra_exclusion_colour[2:4],
+                                      extra_exclusion_colour[4:6]]
+        exclusion_colour_dec = []
+        for colour in hex_exclusion_colour_split:
+            exclusion_colour_dec.append(int(colour, 16))
+        exclusion_cube_origins.append(exclusion_colour_dec)
 
-    logger.info("Successfully completed the candidates listed on the announcement message.")
+    # Now we have all of the required cube origins, time to check our colour against each.
+    for cube_center in exclusion_cube_origins:
+        in_cube = True
+        for i in range(3):
+            dim_min_max = [cube_center[i] - exclusion_range, cube_center[i] + exclusion_range]
+            if not (dim_min_max[0] < colour_dec_split[i] < dim_min_max[1]):
+                in_cube = False
+                break
+        if in_cube:
+            await ctx.send(f"The colour you have selected is too close to that of an admin role or "
+                           f"protected colour.",
+                           delete_after=delete_messages_after)
+            return
 
-    if vote_type_new:
-        announce_message = announce_message + "\nTo vote, send this command to CPSBot in private message: " \
-                                              "`{0}vote_for <first_choice> <second_choice> <last_choice>`. " \
-                                              "Please understand that while the choice names are not case sensitive, " \
-                                              "they __**must be in order and must be spelt the same way, spaces " \
-                                              "split up your votes**__.\n\nHere's an example of its use: " \
-                                              "`{0}vote_for Freddie Kim_Jong-Un Mao`, here you are voting for " \
-                                              "Freddie as your first choice, Lil' Kimmy as your second and Mao as " \
-                                              "your last choice.".format(cmd_prefix)
+    # Not much left to do, only need to create the custom colour role and make sure that it
+    # sits below the lowest defined admin role.
+    admin_role_obj_list = {}
+    for admin_role in admin_role_list:
+        try:
+            admin_role_object = await commands.RoleConverter().convert(ctx, admin_role)
+            admin_role_obj_list[admin_role_object.position] = admin_role_object
+        except discord.ext.commands.errors.BadArgument:
+            logger.info("Admin role defined in config not found in guild.")
 
-    else:
-        announce_message += (
-            f'\nTo vote, send the following command to CPSBot in a __**PRIVATE MESSAGE**__:\n'
-            f'`{cmd_prefix}vote_for {vote_name} <your_choice>`.\n\nWhile <your_choice> is not case '
-            f'sensitive, you should send it exactly the same as in this message to ensure your vote '
-            f'is counted and valid. Understand that you may only vote if you have the role: {vote_right_role}.\n\n'
-            f'The speeches for each candidate have been sent before this message by Freddie.\n\n'
-            f'Your first valid vote will be counted and can not be changed, votes will be made anonymous after an '
-            f'initial validation count to ensure democratic integrity, information regarding which way individuals '
-            f'have voted will be redacted from the democratic integrity report.'
-        )
+    sorted_admin_list_pos = sorted(admin_role_obj_list)
 
-    vote_storage = {"counts": {}, "votes": {}}
+    # Now we have the sorted list of admin roles, let's query all roles and see if we already have
+    # the requested colour created. CPSBot colour roles have the naming convention: CPS[0x<R><G><B>] in hex.
+    try:
+        prev_colour = await commands.RoleConverter().convert(ctx, f"CPS[0x{colour_hex.upper()}]")
+        await prev_colour.edit(position=sorted_admin_list_pos[0])
+        await ctx.author.add_roles(prev_colour, reason="Custom colour requested by DLC Member.")
+        return
+    except commands.BadArgument:
+        # The role doesn't already exist, let's pass.
+        pass
 
-    for candidate in candidate_list:
-        vote_storage["counts"][candidate.lower()] = 0
+    # Now to create the role we wanted all along.
+    new_colour_role = await ctx.guild.create_role(name=f"CPS[0x{colour_hex.upper()}]",
+                                                  reason="Custom colour role generation by CPSBot.",
+                                                  colour=discord.Colour.from_rgb(r=colour_dec_split[0],
+                                                                                 g=colour_dec_split[1],
+                                                                                 b=colour_dec_split[2]))
 
-    logger.info("Set up dictionary for holding vote count information.")
+    await new_colour_role.edit(position=sorted_admin_list_pos[0])
 
-    # Now need to handle directories for vote storage under the vote_name
-    if not os.path.exists("./active_votes/"):
-        os.mkdir("./active_votes/")
-        logger.info("Generated directory for active votes.")
-    if not os.path.exists("./ended_votes/"):
-        os.mkdir("./ended_votes/")
-        logger.info("Generated directory for ended votes.")
+    for invoker_role in ctx.author.roles:
+        if "CPS[0x" in invoker_role.name:
+            await ctx.author.remove_roles(invoker_role, reason="Removing old colour role from user.")
 
-    if vote_type_new:
-        vote_type_str = "freddie_style_vote"
-    else:
-        vote_type_str = "fpbtp_style_vote"
+    await ctx.author.add_roles(new_colour_role, reason="Automatic custom colour allocation by request.")
 
-    with open(f"./active_votes/vote_{vote_name}_{vote_type_str}.json", mode="w") as json_file:
-        json.dump(vote_storage, json_file, indent=4)
-        logger.info(f"Written JSON file to ./active_votes/vote_{vote_name}_{vote_type_str}.json.")
+    await asyncio.sleep(0.5)
+    for role in ctx.guild.roles:
+        if "CPS[0x" in role.name:
+            if not role.members:
+                await role.delete(reason="Automatic custom colour deletion when unused.")
 
-    target_announce_channel = await commands.TextChannelConverter().convert(ctx, announce_channel_config)
-
-    await target_announce_channel.send(announce_message)
-
-
-@bot.command(pass_context=True)
-async def vote_for(ctx, vote_name, *, votes):
-    """Used for special votes configured by bot owner."""
-
-    global vote_file_queue
-    global vote_types
-
-    thread_id = random.randint(1, 10000)
-
-    while thread_id in vote_file_queue:
-        thread_id = random.randint(1, 10000)
-
-    logging.info("Thread with ID: {} has been opened in a list of length: {}.".format(thread_id, len(vote_file_queue)))
-
-    if not vote_file_queue:
-        await vote_file_queue.append(thread_id)
-        await asyncio.sleep(0.5)
-    else:
-        await vote_file_queue.append(thread_id)
-
-    while vote_file_queue[0] != thread_id:
-        await asyncio.sleep(0.1)
-
-    if (len(vote_file_queue) - 1) > 0:
-        logger.info(f"File queue for votes is currently {len(vote_file_queue)} long.")
-
-    found_vote = False
-    for vote_type in vote_types:
-        if f"./active_votes/vote_{vote_name}_{vote_type}.json" in os.listdir("./active_votes/"):
-            found_vote = True
-            break
-
-    if not found_vote:
-        logger.debug("User attempted to vote in a vote that does not exist.")
-        await ctx.send("The <vote_name> you chose does not exist.")
-
-    #with open()
-
-    vote_file_queue.pop(0)
 
 
 # Begin logging
